+++ extension.js
@@ -1,6 +1,11 @@
 const vscode = require("vscode");
-const translate = require("google-translate-query");
 const languages = require("./languages.js");
+const translate = require("google-translate-open-api").default;
+const he = require("he");
+const path = require('path');
+const vscodeLanguageClient = require('vscode-languageclient');
+const humanizeString = require("humanize-string");
+const camelcase = require("camelcase");
 
 /**
  * @typedef TranslateRes
@@ -15,6 +20,8 @@ const languages = require("./languages.js");
  */
 const recentlyUsed = [];
 
+let client = null;
+
 /**
  * Updates languages lists for the convenience of users
  *
@@ -22,17 +29,10 @@ const recentlyUsed = [];
  * @returns {undefined}
  */
 function updateLanguageList(selectedLanguage) {
-  if (recentlyUsed.find(r => r.value === selectedLanguage.value)) {
+  const index = recentlyUsed.findIndex((r) => r === selectedLanguage);
+  if (index !== -1) {
     // Remove the recently used language from the list
-    const index = recentlyUsed.findIndex(
-      r => r.value === selectedLanguage.value
-    );
     recentlyUsed.splice(index, 1);
-  }
-  if (languages.find(r => r.value === selectedLanguage.value)) {
-    // Remove the recently used language from languages list
-    const index = languages.findIndex(r => r.value === selectedLanguage.value);
-    languages.splice(index, 1);
   }
   // Add the language in recently used languages
   recentlyUsed.splice(0, 0, selectedLanguage);
@@ -78,20 +78,57 @@ function getSelectedLineText(document, selection) {
  */
 function getTranslationPromise(selectedText, selectedLanguage, selection) {
   return new Promise((resolve, reject) => {
-    translate(selectedText, { to: selectedLanguage })
-      .then(res => {
-        if (!!res && !!res.text) {
-          resolve(/** @type {TranslateRes} */{
-            selection,
-            translation: res.text
-          });
+    const { host, port, username, password } = getProxyConfig();
+    const translationConfiguration = {
+      to: selectedLanguage,
+    };
+    if (!!host) {
+      translationConfiguration.proxy = {
+        host,
+        port: Number(port),
+      };
+      if (!!username) {
+        translationConfiguration.proxy.auth = {
+          username,
+          password,
+        };
+      }
+    }
+    translate(selectedText, translationConfiguration)
+      .then((res) => {
+        if (!!res && !!res.data) {
+          // If google rejects the string it will return the same string as input
+          // We can try to split the string into parts, then translate again. Then return it to a
+          // camel style casing
+          if (res.data[0] === selectedText) {
+            translate(humanizeString(selectedText), translationConfiguration)
+              .then(res => {
+                if (!!res && !!res.data) {
+                  resolve(
+                    /** @type {TranslateRes} */ {
+                      selection,
+                      translation: camelcase(res.data[0])
+                    }
+                  );
+                } else {
+                  reject(new Error("Google Translation API issue"));
+                }
+              });
+          } else {
+            resolve(
+              /** @type {TranslateRes} */ {
+                selection,
+                translation: res.data[0]
+              }
+            );
+          }
         } else {
           reject(new Error("Google Translation API issue"));
         }
       })
-      .catch(e => {
-        reject(new Error("Google Translation API issue", e));
-      });
+      .catch((e) =>
+        reject(new Error("Google Translation API issue: " + e.message))
+      );
   });
 }
 
@@ -104,7 +141,7 @@ function getTranslationPromise(selectedText, selectedLanguage, selection) {
  * @returns {Array.<Promise<TranslateRes>>}
  */
 function getTranslationsPromiseArray(selections, document, selectedLanguage) {
-  return selections.map(selection => {
+  return selections.map((selection) => {
     const selectedText = getSelectedText(document, selection);
     return getTranslationPromise(selectedText, selectedLanguage, selection);
   });
@@ -123,21 +160,59 @@ function getTranslationsPromiseArrayLine(
   document,
   selectedLanguage
 ) {
-  return selections.map(selection => {
+  return selections.map((selection) => {
     const selectedLineText = getSelectedLineText(document, selection);
     return getTranslationPromise(selectedLineText, selectedLanguage, selection);
   });
 }
 
 /**
- * Returns user settings Preferred language
+ * Returns user settings Preferred language.
+ * If user hasn't set preferred lang. Prompt to set.
+ */
+function getPreferredLanguage() {
+  return (
+    vscode.workspace
+      .getConfiguration("vscodeGoogleTranslate")
+      .get("preferredLanguage") || setPreferredLanguage()
+  );
+}
+
+async function setPreferredLanguage() {
+  const quickPickData = recentlyUsed
+    .map((r) => ({
+      label: r,
+      description: "(recently used)",
+    }))
+    .concat(languages.map((r) => ({ label: r.name })));
+
+  const selectedLanguage = await vscode.window.showQuickPick(quickPickData);
+  if (!selectedLanguage) {
+    return;
+  }
+  vscode.workspace
+    .getConfiguration()
+    .update(
+      "vscodeGoogleTranslate.preferredLanguage",
+      selectedLanguage.label,
+      vscode.ConfigurationTarget.Global
+    );
+  return selectedLanguage.label;
+}
+
+/**
+ * Returns user settings proxy config
  *
  * @returns {string}
  */
-function getPreferredLanguage() {
-  return vscode.workspace
-    .getConfiguration("vscodeGoogleTranslate")
-    .get("preferredLanguage");
+function getProxyConfig() {
+  const config = vscode.workspace.getConfiguration("vscodeGoogleTranslate");
+  return {
+    host: config.get("proxyHost"),
+    port: config.get("proxyPort"),
+    username: config.get("proxyUsername"),
+    password: config.get("proxyPassword"),
+  };
 }
 
 /**
@@ -146,60 +221,66 @@ function getPreferredLanguage() {
  * @param {vscode.ExtensionContext} context
  * @returns {undefined} There is no an API public surface now (7/3/2019)
  */
-function activate(context) {
-  let translateText = vscode.commands.registerCommand(
+async function activate(context) {
+  const translateText = vscode.commands.registerCommand(
     "extension.translateText",
-    function() {
+    function () {
       const editor = vscode.window.activeTextEditor;
       const { document, selections } = editor;
 
       const quickPickData = recentlyUsed
-        .map(r => ({
-          name: r.name.includes("(recently used)")
-            ? r.name
-            : `${r.name} (recently used)`,
-          value: r.value
+        .map((r) => ({
+          label: r,
+          description: "(recently used)",
         }))
-        .concat(languages);
+        .concat(languages.map((r) => ({ label: r.name })));
 
       vscode.window
-        .showQuickPick(quickPickData.map(l => l.name))
-        .then(res => {
-          if (!res) return;
-          const selectedLanguage = quickPickData.find(t => t.name === res);
-          updateLanguageList(selectedLanguage);
+        .showQuickPick(quickPickData)
+        .then((selectedLanguage) => {
+          if (!selectedLanguage) return;
+          updateLanguageList(selectedLanguage.label);
           const translationsPromiseArray = getTranslationsPromiseArray(
             selections,
             document,
-            selectedLanguage.value
+            languages.find((r) => r.name === selectedLanguage.label).value
           );
           Promise.all(translationsPromiseArray)
-            .then(function(results) {
-              editor.edit(builder => {
-                results.forEach(r => {
+            .then(function (results) {
+              editor.edit((builder) => {
+                results.forEach((r) => {
                   if (!!r.translation) {
-                    builder.replace(r.selection, r.translation);
+                    builder.replace(r.selection, he.decode(r.translation));
                   }
                 });
               });
             })
-            .catch(e => vscode.window.showErrorMessage(e));
+            .catch((e) => vscode.window.showErrorMessage(e.message));
         })
-        .catch(err => {
-          vscode.window.showErrorMessage(err);
+        .catch((err) => {
+          vscode.window.showErrorMessage(err.message);
         });
     }
   );
   context.subscriptions.push(translateText);
 
-  let translateTextPreferred = vscode.commands.registerCommand(
+  const setPreferredLanguageFnc = vscode.commands.registerCommand(
+    "extension.setPreferredLanguage",
+    setPreferredLanguage
+  );
+  context.subscriptions.push(setPreferredLanguageFnc);
+
+  const translateTextPreferred = vscode.commands.registerCommand(
     "extension.translateTextPreferred",
-    function() {
+    async function () {
       const editor = vscode.window.activeTextEditor;
       const { document, selections } = editor;
 
       // vscodeTranslate.preferredLanguage
-      let locale = getPreferredLanguage();
+      const preferredLanguage = await getPreferredLanguage();
+      const locale = languages.find(
+        (element) => element.name === preferredLanguage
+      ).value;
       if (!locale) {
         return;
       }
@@ -210,53 +291,53 @@ function activate(context) {
         locale
       );
       Promise.all(translationsPromiseArray)
-        .then(function(results) {
-          editor.edit(builder => {
-            results.forEach(r => {
+        .then(function (results) {
+          editor.edit((builder) => {
+            results.forEach((r) => {
               if (!!r.translation) {
-                builder.replace(r.selection, r.translation);
+                builder.replace(r.selection, he.decode(r.translation));
               }
             });
           });
         })
-        .catch(e => vscode.window.showErrorMessage(e));
+        .catch((e) => vscode.window.showErrorMessage(e.message));
     }
   );
   context.subscriptions.push(translateTextPreferred);
 
-  let translateLinesUnderCursor = vscode.commands.registerCommand(
-    'extension.translateLinesUnderCursor',
+  const translateLinesUnderCursor = vscode.commands.registerCommand(
+    "extension.translateLinesUnderCursor",
     function translateLinesUnderCursorcallback() {
       const editor = vscode.window.activeTextEditor;
       const { document, selections } = editor;
 
       const quickPickData = recentlyUsed
-      .map(r => ({
-        name: r.name.includes("(recently used)")
-          ? r.name
-          : `${r.name} (recently used)`,
-        value: r.value
-      }))
-      .concat(languages);
+        .map((r) => ({
+          label: r.name,
+          description: "(recently used)",
+        }))
+        .concat(languages.map((r) => ({ label: r.name })));
 
       vscode.window
-        .showQuickPick(quickPickData.map(l => l.name))
-        .then(res => {
-          if (!res) return;
-          const selectedLanguage = quickPickData.find(t => t.name === res);
-          updateLanguageList(selectedLanguage);
+        .showQuickPick(quickPickData)
+        .then((selectedLanguage) => {
+          if (!selectedLanguage) return;
+          updateLanguageList(selectedLanguage.label);
           const translationsPromiseArray = getTranslationsPromiseArrayLine(
             selections,
             document,
-            selectedLanguage.value
+            languages.find((r) => r.name === selectedLanguage.label).value
           );
           Promise.all(translationsPromiseArray)
-            .then(function(results) {
-              editor.edit(builder => {
-                results.forEach(r => {
+            .then(function (results) {
+              editor.edit((builder) => {
+                results.forEach((r) => {
                   if (!!r.translation) {
-                    const ffix = ['', '\n'];
-                    if (editor.document.lineCount - 1 === r.selection.start.line)
+                    const ffix = ["", "\n"];
+                    if (
+                      editor.document.lineCount - 1 ===
+                      r.selection.start.line
+                    )
                       [ffix[0], ffix[1]] = [ffix[1], ffix[0]];
                     const p = new vscode.Position(r.selection.start.line + 1);
                     builder.insert(p, `${ffix[0]}${r.translation}${ffix[1]}`);
@@ -264,26 +345,29 @@ function activate(context) {
                 });
               });
             })
-          .catch(e => vscode.window.showErrorMessage(e));
+            .catch((e) => vscode.window.showErrorMessage(e.message));
         })
-        .catch(err => {
-          vscode.window.showErrorMessage(err);
+        .catch((err) => {
+          vscode.window.showErrorMessage(err.message);
         });
     }
   );
 
   context.subscriptions.push(translateLinesUnderCursor);
 
-  let translateLinesUnderCursorPreferred = vscode.commands.registerCommand(
-    'extension.translateLinesUnderCursorPreferred',
-    function translateLinesUnderCursorPreferredcallback() {
+  const translateLinesUnderCursorPreferred = vscode.commands.registerCommand(
+    "extension.translateLinesUnderCursorPreferred",
+    async function translateLinesUnderCursorPreferredcallback() {
       const editor = vscode.window.activeTextEditor;
       const { document, selections } = editor;
-      let locale = getPreferredLanguage();
+      const preferredLanguage = await getPreferredLanguage();
+      const locale = languages.find(
+        (element) => element.name === preferredLanguage
+      ).value;
       if (!locale) {
         vscode.window.showWarningMessage(
-          'Prefered language is requeried for this feature! Please set this in the settings.'
-          );
+          "Prefered language is requeried for this feature! Please set this in the settings."
+        );
         return;
       }
 
@@ -294,11 +378,11 @@ function activate(context) {
       );
 
       Promise.all(translationsPromiseArray)
-        .then(function(results) {
-          editor.edit(builder => {
-            results.forEach(r => {
+        .then(function (results) {
+          editor.edit((builder) => {
+            results.forEach((r) => {
               if (!!r.translation) {
-                const ffix = ['', '\n'];
+                const ffix = ["", "\n"];
                 if (editor.document.lineCount - 1 === r.selection.start.line)
                   [ffix[0], ffix[1]] = [ffix[1], ffix[0]];
                 const p = new vscode.Position(r.selection.start.line + 1);
@@ -307,14 +391,86 @@ function activate(context) {
             });
           });
         })
-        .catch(e => vscode.window.showErrorMessage(e));
+        .catch((e) => vscode.window.showErrorMessage(e.message));
     }
   );
-
   context.subscriptions.push(translateLinesUnderCursorPreferred);
+
+  // Don't initialize the server if it's not wanted
+  if (!vscode.workspace.getConfiguration("vscodeGoogleTranslate").get("HoverTranslations")) {
+    return;
+  }
+
+  // All Below code initializes the Comment Hovering Translation feature
+  let serverModule = context.asAbsolutePath(path.join('server', 'out', 'server.js'));
+  // The debug options for the server
+  // --inspect=6009: runs the server in Node's Inspector mode so VS Code can attach to the server for debugging
+  let debugOptions = { execArgv: ['--nolazy', '--inspect=16009'] };
+  // If the extension is launched in debug mode then the debug server options are used
+  // Otherwise the run options are used
+  let serverOptions = {
+      run: { module: serverModule, transport: vscodeLanguageClient.TransportKind.ipc },
+      debug: {
+          module: serverModule,
+          transport: vscodeLanguageClient.TransportKind.ipc,
+          options: debugOptions
+      }
+  };
+  let extAll = vscode.extensions.all;
+  let languageId = 2;
+  let grammarExtensions = [];
+  let canLanguages = [];
+  extAll.forEach(extension => {
+      if (!(extension.packageJSON.contributes && extension.packageJSON.contributes.grammars))
+          return;
+      let languages = [];
+      (extension.packageJSON.contributes && extension.packageJSON.contributes.languages || []).forEach((language) => {
+          languages.push({
+              id: languageId++,
+              name: language.id
+          });
+      });
+      grammarExtensions.push({
+          languages: languages,
+          value: extension.packageJSON.contributes && extension.packageJSON.contributes.grammars,
+          extensionLocation: extension.extensionPath
+      });
+      canLanguages = canLanguages.concat(extension.packageJSON.contributes.grammars.map((g) => g.language));
+  });
+  let BlackLanguage = ['log', 'Log'];
+  let userLanguage = vscode.env.language;
+  // Options to control the language client
+  let clientOptions = {
+      // Register the server for plain text documents
+      revealOutputChannelOn: 4,
+      initializationOptions: {
+          grammarExtensions, appRoot: vscode.env.appRoot, userLanguage
+      },
+      documentSelector: canLanguages.filter(v => v).filter((v) => BlackLanguage.indexOf(v) < 0),
+  };
+  // Create the language client and start the client.
+  client = new vscodeLanguageClient.LanguageClient('CommentTranslate', 'Comment Translate', serverOptions, clientOptions);
+  // Start the client. This will also launch the server
+  client.start();
+  await client.onReady();
+  client.onRequest('selectionContains', (textDocumentPosition) => {
+      let editor = vscode.window.activeTextEditor;
+      if (editor && editor.document.uri.toString() === textDocumentPosition.textDocument.uri) {
+          let position = new vscode.Position(textDocumentPosition.position.line, textDocumentPosition.position.character);
+          let selection = editor.selections.find((selection) => {
+              return !selection.isEmpty && selection.contains(position);
+          });
+          if (selection) {
+              return {
+                  range: selection,
+                  comment: editor.document.getText(selection)
+              };
+          }
+      }
+      return null;
+  });
 }
 exports.activate = activate;
-
 
 /**
  * Platform binding function
@@ -323,5 +479,10 @@ exports.activate = activate;
  * @param {vscode.ExtensionContext} context
  * @returns {undefined} There is no an API public surface now (7/3/2019)
  */
-function deactivate() {}
+function deactivate() {
+  if (!client) {
+    return undefined;
+  }
+  return client.stop();
+}
 exports.deactivate = deactivate;