+++ CHANGELOG.md
@@ -1,5 +1,47 @@
 # Change Log
 
+## 1.4.9
+- Fix Hebrew language
+
+## 1.4.8
+- Show's translations when hovering over code and text
+
+## 1.4.7
+- Translate camel case support
+
+## 1.4.6
+- Improve default language settings
+- Fix default shortcut
+
+## 1.4.5
+- Update translation library
+
+## 1.4.4
+- Add Uyghur language
+
+## 1.4.3
+- Add Chinese Traditional language
+
+## 1.4.2
+- Improve CI/CD
+
+## 1.4.1
+- Fix multi-line text translation
+
+## 1.4.0
+- Proxy support
+- Fix HTML entity not decoded
+- CI/CD with Github Actions
+
+## 1.3.3
+- Fix error messages not displayed
+
+## 1.3.2
+- Fix 'plugin stopped working' due to google unofficial api that stopped working
+
+## 1.3.1
+- Fix 'translation does nothing' random issue
+
 ## 1.3.0
 - Add the translate of a line feature